+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 # <img src="https://uploads-ssl.webflow.com/5ea5d3315186cf5ec60c3ee4/5edf1c94ce4c859f2b188094_logo.svg" alt="Pip.Services Logo" width="200"> <br/> Remote Procedure Calls for Python Changelog
 
-## <a name="3.2.7"></a> 3.2.7 (2021-09-01)
+## <a name="3.2.7-3.2.9"></a> 3.2.7-3.2.9 (2021-09-07)
+
+# Features
+* Remove netifaces dependency
+* Add verification for supported REST methods
 
 ### Bug Fixes
 * Fixed json params for requests in RestClient
+* Fixed headers and request params passing
+* Fixed query and body params passing
 
 
 ## <a name="3.2.5-3.2.6"></a> 3.2.5-3.2.6 (2021-08-30)